Moved Random Uuid Generator
Moved the random uuid generator so that it can sit along side the
linear uuid generator.
